@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Snppts.Infrastructure;
 
@@ -16,7 +15,7 @@ namespace Snppts.ViewComponents
             _snippets = snippets.ToArray();
         }
 
-        public async Task<IViewComponentResult> InvokeAsync()
+        public IViewComponentResult Invoke()
         {
             var random = new Random();
             var snippets = _snippets.OrderBy(x => random.Next()).Take(3).ToList();
