@@ -1,5 +1,6 @@
 ﻿@using X.PagedList.Mvc.Core;
-@using X.PagedList.Mvc.Common
+@using X.PagedList.Mvc.Core.Common;
+@using X.PagedList.Web.Common
 @using X.PagedList;
 @using Snppts.Extensions;
 
@@ -96,4 +97,3 @@ else
         </div>
     </div>
 }
-
